Revert "Merge pull request #68 from smallcode/patch-1"
Never mind, I do seem to have support for this.  I'm not sure
why @smallcode was having issues with it, but I'm going to
beef up the test for it a bit to try to figure out what's broken.

This reverts commit c83134ca39569890e3e0233ce97c47ac40ba754d, reversing
changes made to b7cbd181f461e9e6a0d83c844e50083218dcfe9f.