Handle request body the same way, whether it comes from file or string. (issue #163)
This fixes the test failure I introduced in 1deb078ad8db. In that case, running

  request.body = somefunc(request.body)

when request.body came from a file caused weird stuff to happen: under
Python 2, request.body ended up being the repr() of a StringIO object.
Looks like repeated assignment to request.body is not a good idea.

The fix is to ensure request.body is always a byte string in the
Request class itself.