Avoid force_reset to make patching thread-safe
Before, force_reset was used twice when instantiating the original
baseclasses. It was necessary to unpatch things, since the classes
were subclasses of patched classes *and* because the subclass referred
to the super class explicitly by name.

That is, httplib has code like:

  class HTTPConnection:
      ...

  class HTTPSConnection(HTTPConnection):
      "This class allows communication via SSL."

      default_port = HTTPS_PORT

      def __init__(self, host, port=None, ...):
          HTTPConnection.__init__(self, host, port, ...)

Creating a real HTTPSConnection instance is now impossible as long as
httplib.HTTPConnection is patched.

To get around this, force_reset was used to temporary unpatch things.
However, this means modifying global state and is thus not
thread-safe. Modifying the classes once when vcr is loaded is better
since no further patching and unpatching needs to be done after that.

This code in this commit uses byteplay, which is not Python 3
compatible.