Add command line utility for editing cassettes
When you have large response or request bodies it is almoust imposible to edit
then directly in YAML file.

To fix that, I created a command line tool, that will open request or response
body for editing in an external editor and when you save you changes, cassette
file will be updated with you changes.

This saves a lot of time and prevents errors when trying to edit large
responses directly in YAML file.

Teststed this on py27,py34,py35, all tests pass.

Added new section in documentaiton about how to use command line tool.